update still wip

+++ app/Http/Controllers/EstateController.php
@@ -6,12 +6,11 @@ namespace App\Http\Controllers;
 use App\Http\Requests\EstateRequest;
 use App\Models\Estate;
 use App\Models\Option;
-use Illuminate\Contracts\View\View;
-use Illuminate\Database\Eloquent\Relations\BelongsToMany;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 use Inertia\Inertia;
 use Inertia\Response as InertiaResponse;
 
@@ -27,7 +26,7 @@ class EstateController extends Controller
         ]);
     }
 
-    public function findOne(int $id, Request $request): InertiaResponse
+    public function view(int $id): InertiaResponse
     {
         $estate = Estate::findOrFail($id);
 
@@ -37,11 +36,19 @@ class EstateController extends Controller
             'estate' => $estate,
             'user' => Auth::user()
         ]);
+        
     }
 
-    public function showOne(int $id, Request $request): InertiaResponse
+    public function showEdit(int $id): InertiaResponse|RedirectResponse
     {
         $estate = Estate::findOrFail($id);
+
+        $policy = Gate::inspect('update', $estate);
+
+        if(!$policy->allowed()) {
+            return to_route('estate', $id)
+                ->with('error', 'Vous n\'etes pas autorisé à éditer cette ressource');
+        }
 
         return Inertia::render('EstateEdit', [
             'estate' => $estate,
@@ -49,11 +56,17 @@ class EstateController extends Controller
         ]);
     }
 
-    public function createEstate(EstateRequest $request): RedirectResponse
+    public function create(EstateRequest $request): RedirectResponse
     {
         $estate = new Estate();
 
-        $requestData = $request->only(['title', 'price', 'surface', 'rooms']);
+        // $test = Gate::inspect('create', $estate);
+
+        // dd($test);
+
+        $this->authorize('create', $estate);
+
+        $requestData = $request->only(['title', 'price', 'surface', 'rooms', 'user_id']);
 
         foreach ($requestData as $key => $value) {
             $estate->{$key} = $value;
@@ -72,6 +85,8 @@ class EstateController extends Controller
     {
         $estate = Estate::findOrFail($id);
 
+        $this->authorize('update', $estate);
+
         $requestData = $request->only(['title', 'price', 'surface', 'rooms']);
 
         foreach ($requestData as $key => $value) {
@@ -88,6 +103,8 @@ class EstateController extends Controller
     public function removeOne(int $id): Response
     {
         $estate = Estate::findOrFail($id);
+
+        $this->authorize('delete', $estate);
 
         $estate->delete();
